backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
email-validator
python-multipart
passlib[bcrypt]
python-jose
python-dotenv
bcrypt